Move index queries to loader
