+++ packages/loader/src/index.ts
@@ -9,6 +9,8 @@
  *
  * @module loader
  */
+export type { BaseQuery } from '@ceramicnetwork/common'
+export type { Connection } from 'graphql-relay'
 export {
   type CreateOptions,
   DocumentLoader,
@@ -16,4 +18,5 @@ export {
   type UpdateDocOptions,
   type UpdateOptions,
 } from './loader.js'
+export type { ConnectionQuery } from './query.js'
 export type { CacheMap, DeterministicKeysCache, DocID, DocumentCache, LoadKey } from './types.js'
